+++ chess/pieces.py
@@ -13,6 +13,38 @@ class Piece:                       #HERENCIA PIEZAS
       return self.__color__
 
 
+   def find_valid_moves(self, row, col, directions, single_step=False):
+       valid_moves = []
+
+       for delta_row, delta_col in directions:
+        current_row, current_col = row + delta_row, col + delta_col
+        
+        while self.is_within_board(current_row, current_col):
+            target_piece = self.__board__.get_piece(current_row, current_col)
+            
+            if target_piece is not None:
+                if target_piece.get_color() != self.get_color():
+                    valid_moves.append((current_row, current_col))  # Capturar pieza contraria
+                break  # Si hay una pieza, se detiene el avance
+                
+            valid_moves.append((current_row, current_col))  # Movimiento vacío válido
+            
+            if single_step:
+                break  # Solo permite un paso si es necesario
+                
+                # Continuar explorando en la misma dirección
+            current_row += delta_row
+            current_col += delta_col
+        
+       return valid_moves
+
+   def is_within_board(self, row, col):
+      return 0 <= row < 8 and 0 <= col < 8
+
+
+
+
+"""
    def calculate_possible_moves(self, row, col, directions, single_step=False):
        possibles = []
        for row_dir, col_dir in directions:
@@ -33,3 +65,4 @@ class Piece:                       #HERENCIA PIEZAS
                next_row += row_dir
                next_col += col_dir
        return possibles
+"""