+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # CHANGELOG
 
+## [0.42] - 13 / 10 / 2024 - Tests
+
+### Changed
+
+- Test alfil, torre, reina, rey, pieces, caballo
+
+## [0.42] - 11 / 10 / 2024 - Arreglos x codeclimate
+
+### Fixed
+
+- Pieces
+
+### Changed
+
+- Tests
+
 ## [0.42] - 10 / 10 / 2024 - Arreglos clase board
 
 ### Fixed